Remove unused TypeScript Galileo SDK fallback code - removed protect.ts and /api/galileo/protect route (unused) - Removed config.ts (unused) - removed galileo npm dependency (no longer needed) - simplify logger.ts to only use Python service
