+++ typescript/chatbot/sample-project-voice-chatbot/web/package.json
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "@elevenlabs/react": "^0.12.3",
-    "galileo": "^1.31.1",
     "next": "^14.2.0",
     "react": "^18.3.0",
     "react-dom": "^18.3.0"
