+++ typescript/chatbot/sample-project-voice-chatbot/web/src/lib/galileo/logger.ts
@@ -1,11 +1,19 @@
-import { GalileoLogger } from "galileo";
-import { type GuardrailResult } from "./protect";
-
 // Server-side only - this module should only be imported in API routes
 
-// Configuration - use Python service if available for proper Protect Status
+// Python service URL (required for Galileo logging)
 const PYTHON_SERVICE_URL = process.env.GALILEO_PYTHON_SERVICE_URL;
-const USE_PYTHON_SERVICE = !!PYTHON_SERVICE_URL;
+
+export interface GuardrailResult {
+  blocked: boolean;
+  status: "triggered" | "not_triggered";
+  overrideMessage?: string;
+  triggeredRules: Array<{
+    metric: string;
+    value: number;
+    threshold: number;
+    operator: string;
+  }>;
+}
 
 export interface ConversationTurnResult {
   logged: boolean;
@@ -15,36 +23,29 @@ export interface ConversationTurnResult {
   overrideMessage?: string;
 }
 
-export async function startGalileoSession(
-  sessionId: string,
-  name?: string
-): Promise<string> {
-  // When using Python service, session is auto-created on first log-conversation-turn call
-  // No need to explicitly start a session here
-  if (USE_PYTHON_SERVICE) {
-    console.log(`[Galileo] Session ${sessionId} will be created by Python service on first log`);
-    return sessionId;
-  }
-
-  // Fallback: TypeScript SDK (requires GALILEO_API_KEY)
-  const logger = new GalileoLogger({
-    projectName: process.env.GALILEO_PROJECT_NAME || "voice-chatbot",
-    logStreamName: process.env.GALILEO_LOG_STREAM || "default",
-  });
-
-  const sessionName = name || `ElevenLabs-${sessionId.slice(0, 8)}`;
-
-  await logger.startSession({
-    name: sessionName,
-    externalId: sessionId,
-  });
-
-  await logger.flush();
-
-  console.log(`Galileo session started: ${sessionId}`);
+/**
+ * Start a Galileo session for tracking conversation turns.
+ * Session is auto-created by the Python service on first log call.
+ * @param sessionId - Unique identifier for the session
+ * @returns Promise resolving to the session ID
+ */
+export async function startGalileoSession(sessionId: string): Promise<string> {
+  // Session is auto-created on first log-conversation-turn call
+  console.log(`[Galileo] Session ${sessionId} will be created by Python service on first log`);
   return sessionId;
 }
 
+/**
+ * Log a conversation turn to Galileo with optional guardrail checks.
+ * @param sessionId - The session ID for this conversation
+ * @param userTranscript - The user's transcribed speech input
+ * @param agentResponse - The agent's response text
+ * @param turnNumber - Sequential turn number in the conversation
+ * @param latencyMs - Response latency in milliseconds
+ * @param conversationContext - Array of conversation messages for context
+ * @param options - Optional configuration (checkGuardrails: enable input/output filtering)
+ * @returns Promise with logging status and guardrail results
+ */
 export async function logConversationTurn(
   sessionId: string,
   userTranscript: string,
@@ -54,145 +55,90 @@ export async function logConversationTurn(
   conversationContext: Array<{ role: string; content: string }>,
   options: { checkGuardrails?: boolean } = {}
 ): Promise<ConversationTurnResult> {
+  if (!PYTHON_SERVICE_URL) {
+    console.error("[Galileo] GALILEO_PYTHON_SERVICE_URL not configured");
+    return { logged: false, blocked: false };
+  }
+
   const { checkGuardrails = false } = options;
   const protectEnabled = process.env.GALILEO_PROTECT_ENABLED === "true";
   const stageId = process.env.GALILEO_PROTECT_STAGE_ID;
   const projectName = process.env.GALILEO_PROJECT_NAME || "voice-chatbot";
 
-  // Use Python service for ALL logging when available
-  // This ensures protect spans are in the same trace as conversation (required for Protect Status)
-  if (USE_PYTHON_SERVICE) {
-    console.log("[Galileo] Using Python service for unified logging with Protect Status");
-    console.log("[Galileo] DEBUG: stageId=", stageId);
-    console.log("[Galileo] DEBUG: protectEnabled=", protectEnabled);
-    console.log("[Galileo] DEBUG: checkGuardrails=", checkGuardrails);
-    console.log("[Galileo] DEBUG: check_guardrails value=", checkGuardrails && protectEnabled);
-    console.log("[Galileo] DEBUG: userTranscript=", userTranscript.substring(0, 50));
+  console.log("[Galileo] Logging turn via Python service");
 
-    try {
-      const response = await fetch(`${PYTHON_SERVICE_URL}/log-conversation-turn`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({
-          session_id: sessionId,
-          user_transcript: userTranscript,
-          agent_response: agentResponse,
-          turn_number: turnNumber,
-          latency_ms: latencyMs,
-          conversation_context: conversationContext,
-          check_guardrails: checkGuardrails && protectEnabled,
-          stage_id: stageId,
-          project_name: projectName,
-        }),
-      });
+  try {
+    const response = await fetch(`${PYTHON_SERVICE_URL}/log-conversation-turn`, {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({
+        session_id: sessionId,
+        user_transcript: userTranscript,
+        agent_response: agentResponse,
+        turn_number: turnNumber,
+        latency_ms: latencyMs,
+        conversation_context: conversationContext,
+        check_guardrails: checkGuardrails && protectEnabled,
+        stage_id: stageId,
+        project_name: projectName,
+      }),
+    });
 
-      if (!response.ok) {
-        throw new Error(`Python service error: ${response.status}`);
-      }
-
-      const result = await response.json();
-
-      const inputGuardrail: GuardrailResult | undefined = result.input_guardrail ? {
-        blocked: result.input_guardrail.blocked,
-        status: result.input_guardrail.status,
-        triggeredRules: result.input_guardrail.triggered_rules || [],
-        overrideMessage: result.input_guardrail.override_message,
-      } : undefined;
-
-      const outputGuardrail: GuardrailResult | undefined = result.output_guardrail ? {
-        blocked: result.output_guardrail.blocked,
-        status: result.output_guardrail.status,
-        triggeredRules: result.output_guardrail.triggered_rules || [],
-        overrideMessage: result.output_guardrail.override_message,
-      } : undefined;
-
-      if (result.blocked) {
-        console.log(`Turn ${turnNumber} blocked by guardrail`);
-      }
-
-      console.log(`Galileo turn ${turnNumber} logged via Python (blocked=${result.blocked})`);
-
-      return {
-        logged: true,
-        inputGuardrail,
-        outputGuardrail,
-        blocked: result.blocked,
-        overrideMessage: result.override_message,
-      };
-    } catch (error) {
-      console.error("Python service logging failed, falling back to TypeScript:", error);
-      // Fall through to TypeScript logging
+    if (!response.ok) {
+      throw new Error(`Python service error: ${response.status}`);
     }
+
+    const result = await response.json();
+
+    const inputGuardrail: GuardrailResult | undefined = result.input_guardrail ? {
+      blocked: result.input_guardrail.blocked,
+      status: result.input_guardrail.status,
+      triggeredRules: result.input_guardrail.triggered_rules || [],
+      overrideMessage: result.input_guardrail.override_message,
+    } : undefined;
+
+    const outputGuardrail: GuardrailResult | undefined = result.output_guardrail ? {
+      blocked: result.output_guardrail.blocked,
+      status: result.output_guardrail.status,
+      triggeredRules: result.output_guardrail.triggered_rules || [],
+      overrideMessage: result.output_guardrail.override_message,
+    } : undefined;
+
+    if (result.blocked) {
+      console.log(`[Galileo] Turn ${turnNumber} blocked by guardrail`);
+    }
+
+    console.log(`[Galileo] Turn ${turnNumber} logged (blocked=${result.blocked})`);
+
+    return {
+      logged: true,
+      inputGuardrail,
+      outputGuardrail,
+      blocked: result.blocked,
+      overrideMessage: result.override_message,
+    };
+  } catch (error) {
+    console.error("[Galileo] Failed to log turn:", error);
+    return { logged: false, blocked: false };
   }
-
-  // Fallback: TypeScript logging (Protect Status will show N/A)
-  console.log("[Galileo] Using TypeScript logging (Protect Status will show N/A)");
-
-  const logger = new GalileoLogger({
-    projectName: projectName,
-    logStreamName: process.env.GALILEO_LOG_STREAM || "default",
-  });
-
-  await logger.startSession({
-    externalId: sessionId,
-  });
-
-  logger.startTrace({
-    input: userTranscript,
-    name: `Turn-${turnNumber}`,
-    metadata: {
-      session_id: sessionId,
-      turn_number: String(turnNumber),
-      source: "elevenlabs-voice",
-    },
-  });
-
-  logger.addLlmSpan({
-    input: JSON.stringify(conversationContext),
-    output: agentResponse,
-    model: "elevenlabs-agent",
-    name: "Agent_Response",
-    metadata: {
-      latency_ms: String(latencyMs),
-    },
-  });
-
-  logger.conclude({ output: agentResponse });
-  await logger.flush();
-
-  console.log(`Galileo turn ${turnNumber} logged via TypeScript`);
-
-  return {
-    logged: true,
-    blocked: false,
-  };
 }
 
+/**
+ * End a Galileo session.
+ * @param sessionId - The session ID to end
+ */
 export async function endGalileoSession(sessionId: string): Promise<void> {
-  // When using Python service, call its end-session endpoint
-  if (USE_PYTHON_SERVICE) {
-    try {
-      await fetch(`${PYTHON_SERVICE_URL}/end-session/${sessionId}`, {
-        method: "POST",
-      });
-      console.log(`[Galileo] Session ${sessionId} ended via Python service`);
-    } catch (error) {
-      console.error("Failed to end session via Python service:", error);
-    }
+  if (!PYTHON_SERVICE_URL) {
+    console.error("[Galileo] GALILEO_PYTHON_SERVICE_URL not configured");
     return;
   }
 
-  // Fallback: TypeScript SDK (requires GALILEO_API_KEY)
-  const logger = new GalileoLogger({
-    projectName: process.env.GALILEO_PROJECT_NAME || "voice-chatbot",
-    logStreamName: process.env.GALILEO_LOG_STREAM || "default",
-  });
-
-  await logger.startSession({
-    externalId: sessionId,
-  });
-
-  await logger.flush();
-  logger.clearSession();
-  console.log(`Galileo session ended: ${sessionId}`);
+  try {
+    await fetch(`${PYTHON_SERVICE_URL}/end-session/${sessionId}`, {
+      method: "POST",
+    });
+    console.log(`[Galileo] Session ${sessionId} ended`);
+  } catch (error) {
+    console.error("[Galileo] Failed to end session:", error);
+  }
 }
